proxy/requirements.txt
websockets==12.0
python-dotenv==1.0.1
certifi==2024.8.30
